refactor: data 패키지 정리
