+++ app/src/main/java/woowacourse/shopping/data/cartItem/CartItemDatabase.kt
@@ -1,11 +1,10 @@
-package woowacourse.shopping.data.db.cartItem
+package woowacourse.shopping.data.cartItem
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import woowacourse.shopping.data.model.CartItemEntity
 
 @Database(
     entities = [CartItemEntity::class],
